refactor: improve container stop

+++ kura/org.eclipse.kura.ai.triton.server/src/main/java/org/eclipse/kura/ai/triton/server/TritonServerContainerManager.java
@@ -107,12 +107,16 @@ public class TritonServerContainerManager implements TritonServerInstanceManager
     }
 
     private synchronized void stopLocalServer() {
+        if (this.containerID.isEmpty()) {
+            return;
+        }
+
         try {
             this.containerOrchestrationService.stopContainer(containerID);
             this.containerOrchestrationService.deleteContainer(containerID);
             this.containerID = "";
         } catch (KuraException e) {
-            logger.error("Can't stop {} won't stop LOL", containerID);
+            logger.error("Can't stop container ID {}. Caused by", containerID, e);
         }
     }
